moved constants file outside of the pages folder
