+++ pages/index.js
@@ -4,7 +4,7 @@ import styles from "@/styles/Home.module.css";
 import { useState, useEffect } from "react";
 import PlayerInfo from "./components/PlayerInfo";
 import SeasonSelector from "./components/SeasonSelector";
-import { SEASONS } from "./components/constants";
+import { SEASONS } from "../constants";
 import MainImage from "./components/MainImage";
 
 const inter = Inter({ subsets: ["latin"] });
